requirements.txt
cherrypy>=11.0.0,<=11.99
librosa>=0.5,<=0.5.99
numpy>=1.12
Pillow>=4.0.0,<=4.99
PyAudio>=0.2
PyOpenGL>=3.1.0,<=3.1.99
PyOpenGL_accelerate
pyttsx3==2.6
requests>=2.18,<=2.99
SpeechRecognition==3.7.1
